api/thesis/:id : refactoring + api delete
