+++ src/app/api/thesis/route.ts
@@ -63,41 +63,3 @@ export async function POST(request: NextRequest) {
     );
   }
 }
-
-export async function PUT(request: NextRequest) {
-  try {
-    const requestData = await request.json();
-    const timestamp = new Date().toISOString();
-
-    const newThesis: Thesis = {
-      id: requestData.id || +new Date(),
-      judul: requestData.judul,
-      abstrak: requestData.abstrak,
-      nim: requestData.nim,
-      tahun: requestData.tahun,
-      updatedAt: timestamp,
-    };
-    const thesisData = await thesisSchema.parseAsync(newThesis);
-    const updatedThesis = await ThesisModel.updateThesis(
-      requestData.id,
-      thesisData,
-    );
-
-    return NextResponse.json({
-      success: true,
-      message: "updated!",
-      data: updatedThesis,
-    });
-  } catch (error) {
-    console.log(error);
-    return NextResponse.json(
-      {
-        success: false,
-        error,
-      },
-      {
-        status: 500,
-      },
-    );
-  }
-}
